Changed fixed effect of model
By adding the trait:cm allows different regression slopes for the trait(venom components) on cm. Helping us understand if each of the traits is being affected by cm.
(-1) allows better interpretation of the model results by preventing cross-parameter comparisons. 
Finally, we set the random effect as being random=~us(trait):species - what this does is that it allows us to obtain covariance across traits that are influenced by the effect of species.  feel in the context of phylogeny as the random effects this makes more sense because the vectors of the random effect are associated with the inv relationship matrix that we get from the phylogeny (rather than the vectors obtained from multiple measurements of data)